comment out uibuilder.js from actions
