+++ modules/actions/index.js
@@ -10,7 +10,7 @@ export * from './functions';
 export * from './settings';
 export * from './rules';
 export * from './grade';
-export * from './uibuilder';
+// export * from './uibuilder';
 export * from './uibuilderN';
 export * from './monitoring';
 export * from './cache';
